Web Stories Embed Block - Moves embed block files to Stories List block directory.
