--- assets/src/story-embed-block/block/block.json
@@ -1,27 +1,0 @@
-{
-  "name": "web-stories/embed",
-  "category": "embed",
-  "attributes": {
-    "url": {
-      "type": "string"
-    },
-    "title": {
-      "type": "string"
-    },
-    "poster": {
-      "type": "string"
-    },
-    "width": {
-      "type": "number",
-      "default": 360
-    },
-    "height": {
-      "type": "number",
-      "default": 600
-    },
-    "align": {
-      "type": "string",
-      "default": "none"
-    }
-  }
-}
