+++ assets/src/web-stories-block/block/block-types/story-embed-block/edit.js
@@ -33,12 +33,12 @@ import { useDispatch, useSelect } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { trackEvent } from '../../tracking';
+import { trackEvent } from '../../../../tracking';
 import EmbedControls from './embedControls';
 import EmbedLoadinng from './embedLoading';
 import EmbedPlaceholder from './embedPlaceholder';
 import EmbedPreview from './embedPreview';
-import { icon } from './';
+import { ReactComponent as icon } from './icon.svg';
 import './edit.css';
 
 const MIN_SIZE = 20;
