+++ assets/src/web-stories-block/block/block-types/story-embed-block/storyPlayer.js
