+++ assets/src/web-stories-block/block/deprecated.js
